requirements.txt
SQLAlchemy>=2.0,<3.0
alembic>=1.13,<2.0
psycopg2-binary>=2.9,<3.0
requests>=2.31,<3.0
zeep>=4.2,<5.0
pandas>=2.1,<3.0
beautifulsoup4>=4.12,<5.0
playwright>=1.42,<2.0
pydantic>=2.6,<3.0
fastapi>=0.109,<1.0
strawberry-graphql[fastapi]>=0.216,<1.0
uvicorn[standard]>=0.24,<1.0
python-dotenv>=1.0,<2.0
